backend/requirements.txt
fastapi
uvicorn
pydantic-settings
openai